+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/app/fragment/TestOriginFragment.kt
@@ -1,0 +1,3 @@
+package me.panpf.androidxkt.test.app.fragment
+
+class TestOriginFragment : android.app.Fragment()