+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/app/fragment/TestSupportFragment.kt
@@ -1,0 +1,3 @@
+package me.panpf.androidxkt.test.app.fragment
+
+class TestSupportFragment : android.support.v4.app.Fragment()