+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/widget/ToastxTest.kt
@@ -19,8 +19,9 @@ package me.panpf.androidxkt.test.widget
 import android.support.test.InstrumentationRegistry
 import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
-import android.view.View
 import me.panpf.androidxkt.test.R
+import me.panpf.androidxkt.test.app.activity.ActivityxTestActivity
+import me.panpf.androidxkt.test.app.activity.ActivityxTestFragmentActivity
 import me.panpf.androidxkt.view.inflateLayout
 import me.panpf.androidxkt.widget.showLongToast
 import me.panpf.androidxkt.widget.showLongToastWithSelf
@@ -34,39 +35,59 @@ import org.junit.runner.RunWith
 class ToastxTest {
 
     @get:Rule
-    val fragmentRuleToast: ActivityTestRule<ToastxTestActivity> = ActivityTestRule<ToastxTestActivity>(ToastxTestActivity::class.java)
+    val activityRule: ActivityTestRule<ActivityxTestActivity> = ActivityTestRule<ActivityxTestActivity>(ActivityxTestActivity::class.java)
+
+    @get:Rule
+    val fragmentActivityRule: ActivityTestRule<ActivityxTestFragmentActivity> = ActivityTestRule<ActivityxTestFragmentActivity>(ActivityxTestFragmentActivity::class.java)
 
     @Test
     fun testContextToast() {
-        fragmentRuleToast.activity.showLongToast("今天是2018年10月18号")
-        fragmentRuleToast.activity.showLongToast("今天是%d年%d月%d号", 2018, 10, 18)
-        fragmentRuleToast.activity.showLongToast(R.string.toast_test)
-        fragmentRuleToast.activity.showLongToast(R.string.toast_test_tp, 2018, 10, 18)
+        val activity = fragmentActivityRule.activity
 
-        fragmentRuleToast.activity.showShortToast("今天是2018年10月18号")
-        fragmentRuleToast.activity.showShortToast("今天是%d年%d月%d号", 2018, 10, 18)
-        fragmentRuleToast.activity.showShortToast(R.string.toast_test)
-        fragmentRuleToast.activity.showShortToast(R.string.toast_test_tp, 2018, 10, 18)
+        activity.showLongToast("今天是2018年10月18号")
+        activity.showLongToast("今天是%d年%d月%d号", 2018, 10, 18)
+        activity.showLongToast(R.string.toast_test)
+        activity.showLongToast(R.string.toast_test_tp, 2018, 10, 18)
+
+        activity.showShortToast("今天是2018年10月18号")
+        activity.showShortToast("今天是%d年%d月%d号", 2018, 10, 18)
+        activity.showShortToast(R.string.toast_test)
+        activity.showShortToast(R.string.toast_test_tp, 2018, 10, 18)
     }
 
     @Test
-    fun testFragmentToast() {
-        val fragment = fragmentRuleToast.activity.supportFragmentManager.findFragmentById(R.id.testAt_frame)
+    fun testSupportFragmentToast() {
+        val fragment = fragmentActivityRule.activity.getFragment()
 
-        fragment?.showLongToast("今天是2018年10月18号")
-        fragment?.showLongToast("今天是%d年%d月%d号", 2018, 10, 18)
-        fragment?.showLongToast(R.string.toast_test)
-        fragment?.showLongToast(R.string.toast_test_tp, 2018, 10, 18)
+        fragment.showLongToast("今天是2018年10月18号")
+        fragment.showLongToast("今天是%d年%d月%d号", 2018, 10, 18)
+        fragment.showLongToast(R.string.toast_test)
+        fragment.showLongToast(R.string.toast_test_tp, 2018, 10, 18)
 
-        fragment?.showShortToast("今天是2018年10月18号")
-        fragment?.showShortToast("今天是%d年%d月%d号", 2018, 10, 18)
-        fragment?.showShortToast(R.string.toast_test)
-        fragment?.showShortToast(R.string.toast_test_tp, 2018, 10, 18)
+        fragment.showShortToast("今天是2018年10月18号")
+        fragment.showShortToast("今天是%d年%d月%d号", 2018, 10, 18)
+        fragment.showShortToast(R.string.toast_test)
+        fragment.showShortToast(R.string.toast_test_tp, 2018, 10, 18)
+    }
+
+    @Test
+    fun testOriginFragmentToast() {
+        val fragment = activityRule.activity.getFragment()
+
+        fragment.showLongToast("今天是2018年10月18号")
+        fragment.showLongToast("今天是%d年%d月%d号", 2018, 10, 18)
+        fragment.showLongToast(R.string.toast_test)
+        fragment.showLongToast(R.string.toast_test_tp, 2018, 10, 18)
+
+        fragment.showShortToast("今天是2018年10月18号")
+        fragment.showShortToast("今天是%d年%d月%d号", 2018, 10, 18)
+        fragment.showShortToast(R.string.toast_test)
+        fragment.showShortToast(R.string.toast_test_tp, 2018, 10, 18)
     }
 
     @Test
     fun testViewToast() {
-        val view = fragmentRuleToast.activity.findViewById<View>(R.id.testAt_frame)
+        val view = fragmentActivityRule.activity.getView()
 
         view.showLongToast("今天是2018年10月18号")
         view.showLongToast("今天是%d年%d月%d号", 2018, 10, 18)