+++ androidx/src/androidTest/java/me/panpf/androidx/test/app/fragment/TestOriginFragment.java
@@ -1,0 +1,5 @@
+package me.panpf.androidx.test.app.fragment;
+
+public class TestOriginFragment extends android.app.Fragment {
+
+}