+++ androidx/src/androidTest/java/me/panpf/androidx/test/app/fragment/TestSupportFragment.java
@@ -1,0 +1,5 @@
+package me.panpf.androidx.test.app.fragment;
+
+public class TestSupportFragment extends android.support.v4.app.Fragment {
+
+}