+++ androidx/src/androidTest/java/me/panpf/androidx/test/hardware/HardwarexTest.java
@@ -67,6 +67,12 @@ public class HardwarexTest {
     }
 
     @Test
+    public final void testPhoneNumber() {
+        Context context = InstrumentationRegistry.getContext();
+        Assert.assertNotNull(Hardwarex.getPhoneNumber(context));
+    }
+
+    @Test
     public final void testDeviceId() {
         Context context = InstrumentationRegistry.getContext();
         if (Androidx.isAtLeast6_0()) {
@@ -145,6 +151,34 @@ public class HardwarexTest {
         } else {
             String serial = Hardwarex.getSerial();
             Assert.assertTrue("serial: " + serial, Stringx.isSafe(serial));
+        }
+    }
+
+    @Test
+    public final void testIMEI() {
+        Context context = InstrumentationRegistry.getContext();
+        if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
+            String imei = Hardwarex.getIMEI(context);
+            Assert.assertTrue("imei: " + imei, Stringx.isSafe(imei)
+                    && !Comparisonx.areEqual(imei, "unknown")
+                    && !Comparisonx.areEqual(imei, "PermissionDenied"));
+        } else {
+            String imei = Hardwarex.getIMEI(context);
+            Assert.assertEquals("imei: " + imei, imei, "PermissionDenied");
+        }
+    }
+
+    @Test
+    public final void testIMSI() {
+        Context context = InstrumentationRegistry.getContext();
+        if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
+            String imsi = Hardwarex.getIMSI(context);
+            Assert.assertTrue("imsi: " + imsi, Stringx.isSafe(imsi)
+                    && !Comparisonx.areEqual(imsi, "unknown")
+                    && !Comparisonx.areEqual(imsi, "PermissionDenied"));
+        } else {
+            String imsi = Hardwarex.getIMSI(context);
+            Assert.assertEquals("imsi: " + imsi, imsi, "PermissionDenied");
         }
     }